generalization of AMG preconditioner
