+++ include/exadg/structure/spatial_discretization/operator.cpp
@@ -27,7 +27,6 @@
 #include <exadg/solvers_and_preconditioners/preconditioners/jacobi_preconditioner.h>
 #include <exadg/solvers_and_preconditioners/preconditioners/preconditioner_amg.h>
 #include <exadg/solvers_and_preconditioners/solvers/iterative_solvers_dealii_wrapper.h>
-#include <exadg/structure/preconditioners/amg_preconditioner.h>
 #include <exadg/structure/preconditioners/multigrid_preconditioner.h>
 #include <exadg/structure/spatial_discretization/operator.h>
